+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ## [Unreleased]
+## [0.1.1] - 2021-03-25
+### Added
+- replace() method to replace existing services
+- set() method to add/replace services
+- has() method to check if a service is registered
+
 ## 0.1.0 - 2021-02-19
 ### Added
 - Initial version
 
-[Unreleased]: https://github.com/f3ath/klizma/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/f3ath/klizma/compare/0.1.1...HEAD
+[0.1.1]: https://github.com/f3ath/klizma/compare/0.1.0...0.1.1